backend/requirements.txt
Django
djangorestframework
deepseek
openai
djangorestframework-simplejwt
python-dotenv
Faker
django-filter
psycopg2-binary  # For PostgreSQL database connectivity